extend LogicalOrToBooleanOrRector to AND and rename to LogicalToBooleanRector
