+++ config/level/celebrity/celebrity.yaml
@@ -13,4 +13,5 @@ services:
     Rector\Php\Rector\Property\VarToPublicPropertyRector: ~
 
     #  false or true → false || true
-    Rector\Celebrity\Rector\LogicalOr\LogicalOrToBooleanOrRector: ~
+    #  false and true → false && true
+    Rector\Celebrity\Rector\BooleanOp\LogicalToBooleanRector: ~
