+++ packages/Celebrity/tests/Rector/BooleanOp/LogicalToBooleanRector/Fixture/and.php.inc
@@ -1,0 +1,23 @@
+<?php
+
+if ($f = false and true) {
+    return $f;
+}
+
+if (false and true) {
+    return $f;
+}
+
+?>
+-----
+<?php
+
+if (($f = false) && true) {
+    return $f;
+}
+
+if (false && true) {
+    return $f;
+}
+
+?>
