+++ packages/Celebrity/tests/Rector/BooleanOp/LogicalToBooleanRector/Fixture/or.php.inc
